Gen.mutate() rebuild
rewrited code for mutate. It works on interval 0.0 to 1.0 (raw values). It's questionable if it should constrain range before random picking or after. When value is 0 it's 50% chance to be again 0 (it pick random number in range 0-rate  to 0+rate and then constrains to 0-1). Older implementation picked random number in range 0 to 0+range directly).